initial: simplemente una idea de algunos cambios. pd: Es necesario actualizar el modelo
